[feature][CAR-32] Update store management
