+++ frontend/src/app/store/car-filters/facades/car-filters-facade.service.ts
@@ -1,18 +1,11 @@
 import { Injectable, Signal, inject } from '@angular/core';
-import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs';
-import { AppState } from '../../..';
 import { CarBrandDto as CarBrand } from '../../../api/models/car-brand-dto';
 import { CarModelDto as CarModel } from '../../../api/models/car-model-dto';
-import { FeatureKeys } from '../../../constants/feature-keys';
 import { CarFilters } from '../../../tabs/car-filters/models/car-filters.model';
-import * as CarFiltersActions from '../actions/car-filters.actions';
-import * as CarFiltersSelectors from '../selectors/car-filters.selectors';
-import { CarFiltersStore } from '../signals/car-filter-store';
+import { CarFiltersStore } from '../car-filter-store';
 
 @Injectable({ providedIn: 'root' })
 export class CarFiltersFacadeService {
-    private _store = inject(Store<AppState[FeatureKeys.CAR_FILTERS]>);
     private _carFiltersStore = inject(CarFiltersStore);
 
     //Selectors BEGIN -------------------------
@@ -25,15 +18,11 @@ export class CarFiltersFacadeService {
     }
 
     selectCarFiltersResultCount(): Signal<number | undefined> {
-        return this._store.selectSignal(CarFiltersSelectors.selectCarFiltersResultCount);
+        return this._carFiltersStore.resultCount;
     }
 
     selectSelectedCarFilters(): Signal<CarFilters> {
-        return this._store.selectSignal(CarFiltersSelectors.selectSelectedCarFilters);
-    }
-
-    selectAreCarBrandsLoaded(): Observable<boolean> {
-        return this._store.select(CarFiltersSelectors.selectAreCarBrandsLoaded);
+        return this._carFiltersStore.selectedCarFilters;
     }
     //Selectors END -------------------------
 
@@ -47,13 +36,7 @@ export class CarFiltersFacadeService {
     }
 
     getCarFiltersResultCount(selectedCarFiltersQuery: CarFilters): void {
-        this._store.dispatch(
-            CarFiltersActions.getCarFiltersResultCount({ selectedCarFiltersQuery }),
-        );
-    }
-
-    setAreCarBrandsLoaded(areCarBrandsLoaded: boolean): void {
-        this._store.dispatch(CarFiltersActions.setAreCarBrandsLoaded({ areCarBrandsLoaded }));
+        this._carFiltersStore.getCarFiltersResultCount(selectedCarFiltersQuery);
     }
     //Actions END -------------------------
 }
